dexc-desktop: move C code from main.go to app.go
